Consolidate Playwright tests under tests/test_e2e/
Pytest collects (imports) every test module before applying marker
filters, so the unit-test job's `-m "not e2e"` couldn't avoid the
ModuleNotFoundError when Playwright wasn't installed. Move all
Playwright tests under one folder and route by path:

- Rename tests/test_browser → tests/test_e2e
- Move the env_config UI download test from tests/test_env_config_export.py
  into tests/test_e2e/test_env_config_browser.py
- Drop the requires-python=>=3.14 pin from the env_config fixture
  notebook (CI runs Python 3.11; suspected cause of the e2e timeout)
- Workflow: unit job switches to `--ignore=tests/test_e2e`; the previous
  browser-tests and e2e-tests jobs collapse into a single e2e-tests job
  that runs `pytest tests/test_e2e -v`
