+++ tests/fixtures/envconfig_export_notebook.py
@@ -1,5 +1,4 @@
 # /// script
-# requires-python = ">=3.14"
 # dependencies = [
 #     "anywidget==0.11.0",
 #     "marimo>=0.23.3",
